사용자 조회(멀티건) 카운트

+++ src/main/java/egovframework/com/cop/com/service/impl/EgovUserInfManageDAO.java
@@ -45,7 +45,7 @@ public class EgovUserInfManageDAO extends EgovComAbstractDAO {
 	 * @throws Exception
 	 */
 	public int selectUserListCnt(UserInfVO userVO) throws Exception {
-		return (Integer) selectOne("EgovUserInfManageDAO.selectUserListCnt", userVO);
+		return selectOne("EgovUserInfManageDAO.selectUserListCnt", userVO);
 	}
 
 	/**
